+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Adds Content-Type application/json to payload headers.
+- Adds getBody function to get the body from a message.
 - Adds getTag function to get the tag from a message.
 - Adds getSubject function to get the subject from a message.
 - Adds tests to ensure that the json payload has the proper values.